drm/virtio: populate .set_busid callback
Earlier commit was removing all the users of drm_platform_set_busid and
removed the virtio hunk (which uses the PCI version of the function) by
mistake.

This in itself resulted in user space receiving an incorrect value for
the busid, which in the case below lead to the failure to detect
the (correct) device and ultimately the X server failing to start.

Fixes: a325725633c ("drm: Lobotomize set_busid nonsense for !pci
drivers")
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Cc: Laszlo Ersek <lersek@redhat.com>
Cc: stable@vger.kernel.org
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1366842
Reported-by: Laszlo Ersek <lersek@redhat.com>
Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>
---
Since I'm not 100% sure if we can get into .set_busid() as pci_dev is
NULL (for the virtio-vga 'vs' virgio-gpu-pci case) I've left the local
wrapper.

Note: patch is against mainline (refs/tags/v4.8) but should apply for
drm-next/others.